# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CMS manager",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost/cms-template/public",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/admin/product-conversations": {
            "get": {
                "tags": [
                    "Admin Product Conversations"
                ],
                "summary": "Get all product conversations",
                "operationId": "c32c8095ae186fc2e839f1ee81215972",
                "parameters": [
                    {
                        "name": "is_approved",
                        "in": "query",
                        "description": "Filter by approval status",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of product conversations"
                    }
                }
            }
        },
        "/api/admin/product-conversations/{id}": {
            "get": {
                "tags": [
                    "Admin Product Conversations"
                ],
                "summary": "Get a specific product conversation",
                "operationId": "a8cedd347e964065c8869aa0feb2747f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the conversation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product conversation details"
                    },
                    "404": {
                        "description": "Conversation not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin Product Conversations"
                ],
                "summary": "Delete a product conversation",
                "operationId": "1d44994b31ca6000520d7792da617eae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the conversation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conversation deleted"
                    }
                }
            }
        },
        "/api/admin/product-conversations/product/{productId}": {
            "get": {
                "tags": [
                    "Admin Product Conversations"
                ],
                "summary": "Get product conversations by product ID",
                "operationId": "887872742abb2bc0600e3c06ec610f2d",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "The ID of the product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "is_approved",
                        "in": "query",
                        "description": "Filter by approval status",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of product conversations for the specified product"
                    }
                }
            }
        },
        "/api/admin/product-conversations/{conversationId}/replies": {
            "get": {
                "tags": [
                    "Admin Product Conversations"
                ],
                "summary": "Get replies for a conversation",
                "operationId": "eebcb6df5486c2ad41b98a27742b4992",
                "parameters": [
                    {
                        "name": "conversationId",
                        "in": "path",
                        "description": "The ID of the conversation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "is_approved",
                        "in": "query",
                        "description": "Filter by approval status",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of replies for the specified conversation"
                    }
                }
            }
        },
        "/api/admin/product-conversations/{id}/approval": {
            "put": {
                "tags": [
                    "Admin Product Conversations"
                ],
                "summary": "Approve a product conversation",
                "operationId": "594ea80f44ccfa8f2db6a069ef880726",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the conversation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conversation approved"
                    }
                }
            }
        },
        "/api/admin/product-conversations/{id}/admin-reply": {
            "post": {
                "tags": [
                    "Admin Product Conversations"
                ],
                "summary": "Add an admin reply to a conversation",
                "operationId": "81eb3d7fd64ed5dec931f67f0d62ab54",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the conversation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "text"
                                ],
                                "properties": {
                                    "text": {
                                        "description": "Reply text",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Admin reply added"
                    }
                }
            }
        },
        "/api/admin/product-conversations/{conversationId}/admin-reply/{replyId}": {
            "delete": {
                "tags": [
                    "Admin Product Conversations"
                ],
                "summary": "Delete an admin reply from a conversation",
                "operationId": "b2cfd86c6028f83350f55b390c4f2cc9",
                "parameters": [
                    {
                        "name": "conversationId",
                        "in": "path",
                        "description": "The ID of the conversation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "replyId",
                        "in": "path",
                        "description": "The ID of the reply",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Admin reply deleted"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Authentication using Bearer token (Sanctum)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "cookieSession": {
                "type": "apiKey",
                "description": "Authentication using session cookie (laravel_session)",
                "name": "laravel_session",
                "in": "cookie"
            },
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ],
...
}